flask/requirements.txt
Flask==2.0.2
Flask_Cors==3.0.10
google_api_python_client==2.81.0
google_auth_oauthlib==1.0.0
protobuf==4.22.1
